refactor(confirm): convert component to typescript
