+++ src/components/confirm/confirm.config.ts
