+++ src/components/confirm/confirm.style.ts
@@ -1,7 +1,8 @@
 import styled, { css } from "styled-components";
-import PropTypes from "prop-types";
+
 import StyledIcon from "../icon/icon.style";
 import { StyledHeadingTitle } from "../heading/heading.style";
+import { ConfirmProps } from ".";
 
 export const StyledConfirmButtons = styled.div`
   margin-top: 48px;
@@ -9,7 +10,9 @@ export const StyledConfirmButtons = styled.div`
   justify-content: flex-end;
 `;
 
-export const StyledConfirmHeading = styled.div`
+export const StyledConfirmHeading = styled.div<{
+  type: ConfirmProps["iconType"];
+}>`
   display: flex;
   align-items: center;
 
@@ -32,7 +35,3 @@ export const StyledConfirmHeading = styled.div`
       `}
   }
 `;
-
-StyledConfirmHeading.propTypes = {
-  type: PropTypes.oneOf(["error", "warning"]),
-};
