--- src/components/confirm/index.js
@@ -1,1 +1,0 @@
-export { default } from "./confirm.component";
