+++ src/components/confirm/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./confirm.component";
+export type { ConfirmProps } from "./confirm.component";
